validando formularios

+++ app/Http/Controllers/SucursalController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use App\Models\Sucursal;
 use App\Models\Producto;
 use Illuminate\Http\Request;
@@ -24,8 +25,15 @@ class SucursalController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
-            'nombre' => 'required',
-            'direccion' => 'required',
+            'nombre' => 'required|string|max:255|regex:/^[a-zA-ZáéíóúÁÉÍÓÚüÜñÑ\s]+$/',
+            'direccion' => 'required|string|max:255',
+
+        ],[
+            'nombre.regex' => 'El formato del campo nombre no es válido.',
+            'direccion.required' => 'El campo dirección es obligatorio.',
+            'nombre.required' => 'El campo nombre es obligatorio.',
+            'direccion.string' => 'El campo dirección debe ser una cadena de texto.',
+            'direccion.max' => 'El campo dirección no debe tener más de :max caracteres.',
         ]);
 
         $sucursal = new Sucursal();
@@ -41,6 +49,7 @@ class SucursalController extends Controller
         $sucursal = Sucursal::with('productoSucursal')->find($id);
         $productos = Producto::all();
         $productosAsociados = $sucursal->productoSucursal;
+        
 
         return view('sucursales.edit', compact('sucursal', 'productos', 'productosAsociados'));
     }
@@ -49,8 +58,15 @@ class SucursalController extends Controller
     public function update(Request $request, $id)
     {
         $this->validate($request, [
-            'nombre' => 'required',
-            'direccion' => 'required'
+            'nombre' => 'required|string|max:255|regex:/^[a-zA-ZáéíóúÁÉÍÓÚüÜñÑ\s]+$/',
+            'direccion' => 'required|string|max:255',
+
+        ],[
+            'nombre.regex' => 'El formato del campo nombre no es válido.',
+            'direccion.required' => 'El campo dirección es obligatorio.',
+            'nombre.required' => 'El campo nombre es obligatorio.',
+            'direccion.string' => 'El campo dirección debe ser una cadena de texto.',
+            'direccion.max' => 'El campo dirección no debe tener más de :max caracteres.',
         ]);
 
         $sucursal = Sucursal::findOrFail($id);
@@ -64,7 +80,10 @@ class SucursalController extends Controller
     public function destroy($id)
     {
         $sucursal = Sucursal::findOrFail($id);
-        $sucursal->productoSucursal()->delete();
+        if ($sucursal->productoSucursal()->exists()) {
+            return redirect()->route('sucursales.index')->with('error', 'No puedes eliminar una sucursal con productos asociados.');
+        }
+
         $sucursal->delete();
 
         return redirect()->route('sucursales.index')->with('success', 'Sucursal eliminada exitosamente');
